src/main/java/com/weixin/order/sell/service/impl/ProductCategoryServiceImpl.java
package com.weixin.order.sell.service.impl;

import com.weixin.order.sell.dataobject.ProductCategory;
import com.weixin.order.sell.repository.ProductCategoryRepository;
import com.weixin.order.sell.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.awt.print.Pageable;
import java.util.Arrays;
import java.util.List;

/**
 * @Author : yangxudong
 * @Description : 商品类目的service实现类
 * @Date : 下午1:54 2018/5/8
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService{

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }

    @Override
    public List<ProductCategory> findByCategoryTypeList(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }
}
